test(libgen): consolidate test files to one per source file
